backend/requirements.txt

fastapi
uvicorn
langchain
faiss-cpu
pydantic
python-multipart
sentence-transformers
ollama
pypdf
python-docx
sqlalchemy
